refactor: move cover fetching logic from main to renderer process
Addresses PR feedback by moving the API fetch call into the renderer.
This removes the need for IPC communication and allows animated artworks
to work in web-only environments.
