--- src/main/features/core/animated-covers/constants.ts
@@ -1,3 +1,0 @@
-export const ANIMATED_COVERS_API_BASE = 'https://artwork.m8tec.top';
-export const ANIMATED_COVERS_API_ENDPOINT = '/api/v1/artwork/search';
-export const REQUEST_TIMEOUT = 10000;
