--- src/main/features/core/animated-covers/types.ts
@@ -1,6 +1,0 @@
-export interface AnimatedCoverResponse {
-    album?: string;
-    artist?: string;
-    isCached?: boolean;
-    url?: string;
-}
