+++ src/main/features/core/index.ts
@@ -1,4 +1,3 @@
-import './animated-covers';
 import './autodiscover';
 import './lyrics';
 import './player';
